+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): Image loading.
 - [@mjkuranda](https://github.com/mjkuranda): Rendering image for player, asteroid and background.
+- [@mjkuranda](https://github.com/mjkuranda): Distance to player stats.
 
 ## [0.10.0] - 2023-07-11
 ### Added